+++ src/main/java/internet/shop/controllers/shoppingCart/CreateShopCartController.java
@@ -1,4 +1,4 @@
-package internet.shop.controllers;
+package internet.shop.controllers.shoppingCart;
 
 import internet.shop.lib.Injector;
 import internet.shop.model.ShoppingCart;
