+++ src/main/webapp/WEB-INF/views/index.jsp
@@ -5,18 +5,23 @@
 </head>
 <body>
 <h1>Main Page</h1>
+<h3>User Options</h3>
 <button type="button"
         onclick='location.href="http://localhost:8080/registration"'>Registration
 </button>
 <button type="button"
+        onclick='location.href="http://localhost:8080/products/all"'>All Products for User
+</button>
+<br>
+<h3>Admin Options</h3>
+<button type="button"
         onclick='location.href="http://localhost:8080/users/all"'>Get All Users
 </button>
 <button type="button"
-        onclick='location.href="http://localhost:8080/products/all"'>Get All Products
+        onclick='location.href="http://localhost:8080/products/adminAll"'>All Products for Admin
 </button>
 <br>
 <h3>Generators</h3>
-<br>
 <button type="button"
         onclick='location.href="http://localhost:8080/generateUsers"'>Generate Users
 </button>
