+++ src/main/webapp/WEB-INF/views/order.jsp
@@ -2,14 +2,14 @@
 <%@ page contentType="text/html;charset=UTF-8" language="java" %>
 <html>
 <head>
-    <title>allProducts</title>
+    <title>Order</title>
 </head>
 <body>
 <button type="button"
         onclick='location.href="http://localhost:8080/"'>Home
 </button>
 <button type="button"
-        onclick='location.href="http://localhost:8080/shoppingCart"'>Shopping Cart
+        onclick='location.href="http://localhost:8080/products/all"'>Add Product
 </button>
 <table border="1">
     <tr>
@@ -17,7 +17,7 @@
         <th>Name</th>
         <th>Price</th>
     </tr>
-    <c:forEach var="product" items="${products}">
+    <c:forEach var="product" items="${shoppingCartItems}">
         <tr>
             <td>
                 <c:out value="${product.productId}"/>
@@ -29,7 +29,7 @@
                 <c:out value="${product.price}"/>
             </td>
             <td>
-                <a href="${pageContext.request.contextPath}/addProductToShopCart?productId=${product.productId}">Add to Shopping Cart</a>
+                <a href="${pageContext.request.contextPath}/removeProductFromShopCart?productId=${product.productId}">Delete from Shopping Cart</a>
             </td>
         </tr>
     </c:forEach>
