+++ src/main/webapp/WEB-INF/views/products/adminAll.jsp
@@ -2,14 +2,14 @@
 <%@ page contentType="text/html;charset=UTF-8" language="java" %>
 <html>
 <head>
-    <title>allProducts</title>
+    <title>adminAll</title>
 </head>
 <body>
 <button type="button"
         onclick='location.href="http://localhost:8080/"'>Home
 </button>
 <button type="button"
-        onclick='location.href="http://localhost:8080/shoppingCart"'>Shopping Cart
+        onclick='location.href="http://localhost:8080/createProduct"'>Create Product
 </button>
 <table border="1">
     <tr>
@@ -29,7 +29,7 @@
                 <c:out value="${product.price}"/>
             </td>
             <td>
-                <a href="${pageContext.request.contextPath}/addProductToShopCart?productId=${product.productId}">Add to Shopping Cart</a>
+                <a href="${pageContext.request.contextPath}/deleteProduct?productId=${product.productId}">Delete</a>
             </td>
         </tr>
     </c:forEach>
